pytorch/requirements.txt
torch>=1.12
torchvision
numpy
Pillow
tqdm
tensorboard
matplotlib
